pyproject.toml

[tool.poetry]
name = "mamba-music-generation"
version = "0.1.0"
description = ""
authors = ["Jan Retkowski, Milosz Lopatto, Jakub Stepniak"]

[tool.poetry.dependencies]
python = "^3.8"
hydra-core = "^1.1.1"
setuptools = "^69.5.1"
wheel = "^0.43.0"
pytorch-lightning = "^2.2.4"
torch = "2.1.1"
miditok = "^3.0.3"
wandb = "^0.17.0"
muspy = "^0.5.0"
causal-conv1d = "==1.2.*"
torchvision = "0.16.1"
mamba-ssm = "1.2.0.post1"

[tool.poetry.dev-dependencies]
pdoc3 = "^0.10.0"
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
select = ["B", "C", "E", "F", "W", "B9", "I", "Q"]

[tool.ruff.lint.mccabe]
max-complexity = 10
